Add on-by-default `fast` crate feature
Disabling the feature reduces overall code size at the cost of
performance, which is useful for e.g. embedded users.

This feature transitively enables the `basepoint-tables` feature in
`curve25519-dalek`.

The newly introduced `EdwardsPoint::mul_base` API provides a common
abstraction for multiplying by the basepoint which uses the basepoint
tables if available, or falls back to variable base multiplication if
the `fast` feature has been disabled.

This API is now used for both signing and verification.